Handle ISY version 5+ firmware events.
ISY version 5+ firmware adds some additional information to the
events that are sent through the subscripton channel. Specifically
the <action> tag now includes attributes that specify the unit of measure
and precision of the action value.

Prior to version 5, the action tag would look like:

<action>255</action>

Version 5+ firmware will now send:

<action uom="100" prec="0">255</action>

The et2d function is not able to parse this change and results in
setting the action key in the dictionary to the attribute list instead
of the value.

This change makes all the dictionary tag name keys get the value or value
list and restricts the attributes to tag_name-attribute_name keys.